Move Data API into Spin API.
