+++ src/main.cpp
@@ -28,7 +28,6 @@
  */
 
 //--------------OWNTECH APIs----------------------------------
-#include "DataAPI.h"
 #include "TaskAPI.h"
 #include "TwistAPI.h"
 #include "SpinAPI.h"
