--- zephyr/modules/owntech_data_api/zephyr/Kconfig
@@ -1,13 +1,0 @@
-config OWNTECH_DATA_API
-	bool "Enable OwnTech data acquisition using ADCs"
-	default y
-	select DMA
-	depends on CONSOLE_GETCHAR
-	depends on OWNTECH_SPIN_API
-	depends on OWNTECH_FLASH
-	help
-		The data API provides a framework for automatically gathering
-		data from Spin board ADCs into buffers which can be auto generated.
-		It also interfaces with the Twist shield, providing	ways to
-		interact with the sensors available on the shield and converting
-		raw data in their relevant unit.
