+++ zephyr/modules/owntech_spin_api/zephyr/public_api/data_api_internal.h
