+++ zephyr/modules/owntech_spin_api/zephyr/src/data/data_conversion.cpp
@@ -29,7 +29,7 @@
 // Include
 
 // OwnTech API
-#include "DataAPI.h"
+#include "../DataAPI.h"
 
 // Current file header
 #include "data_conversion.h"
