+++ zephyr/modules/owntech_spin_api/zephyr/src/data/data_conversion.h
